remoteloop: drop O_DIRECT
Appart from giving us a hard time on s390x, this feature did not seem
to have a measurable effect. Moreover, O_DIRECT is not supported by
tmpfs so without this patch we could not use tmpfs as backing store,
which does speed up image generation considerably.

Drop the flag and and rather put the store on tmpfs in order to speed
things up.