docs: refactor README
With the introduction of man-pages, this now moves the usage
information, examples, and pipeline descriptions into the man-page. With
this out of the way, the README is restructured with information on how
to install the project, what dependencies exist, and directions on
where to find further information or contact upstream.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>