pipeline: no auto-commit for build stages
Refactor the building of stages and the build tree so that no auto
commit is done at the end of the build pipeline anymore, i.e. the
respective build tree(s) are not commit to the store unless that
was explicitly enabled via a checkpoint.

NB: `objectstore.Object`s are used not via a context manager
anymore, because they are returned from the `build_stages` method
to make the code easier to use and read. Cleanup of Objects during
a KeyboardInterrupt exception (Ctrl-C) are handled by using the
ObjectStore with a context manager, which on exit of the context
will cleanup all objects. Due to a big in python[1] this is indeed
more robust than using `with object_store.new() as tree` because
that is translated[2] to something like:

     1: mgr = (EXPR)
     2: exit = type(mgr).__exit__
     3: value = type(mgr).__enter__(mgr)
 ->  4: # NOTE: KeyboardInterrupt here will "leak" value
     5: try:
     6:	[...]
     7: finally:
     8:	   if exc:
     9:        exit(mgr, None, None, None)

Which can leave the tree initialized but not cleaned up if the
KeyboardInterrupt happens exactly line 4.

[1] https://bugs.python.org/issue29988
[2] https://www.python.org/dev/peps/pep-0343/