docs: fix incorrect option-list formatting
The used format of `-X, --long=VALUE` is not a valid option-list entry,
even though it is very commonly used all over the linux man-pages. Use
the supported format of `-X VALUE, --long=VALUE`, which will format
correctly in the man-page and html outputs.

For reference, these formats are valid in RST option-lists:

        -a                      Short option
        -c arg                  Short option with arg.
        --long                  Long option.
        -2, --two               Aliases on a single line.
        -f FILE, --file=FILE    Aliases with arguments.
        /V                      VMS/DOS-style option.